backendcore/backend/requirements.txt
django==3.2.4
djangorestframework
pillow
